+++ crates/web-client/src/models/account_storage_requirements.rs
@@ -29,6 +29,7 @@ impl SlotAndKeys {
     }
 }
 
+#[derive(Clone)]
 #[wasm_bindgen]
 pub struct AccountStorageRequirements(NativeAccountStorageRequirements);