+++ crates/web-client/src/models/transaction_request/note_id_and_args.rs
@@ -2,6 +2,7 @@ use miden_client::note::NoteId as NativeNoteId;
 use miden_client::transaction::NoteArgs as NativeNoteArgs;
 use wasm_bindgen::prelude::*;
 
+use crate::models::miden_arrays::NoteIdAndArgsArray;
 use crate::models::note_id::NoteId;
 use crate::models::transaction_request::note_and_args::NoteArgs;
 
@@ -37,31 +38,14 @@ impl From<&NoteIdAndArgs> for (NativeNoteId, Option<NativeNoteArgs>) {
     }
 }
 
-#[derive(Clone)]
-#[wasm_bindgen]
-pub struct NoteIdAndArgsArray(Vec<NoteIdAndArgs>);
-
-#[wasm_bindgen]
-impl NoteIdAndArgsArray {
-    #[wasm_bindgen(constructor)]
-    pub fn new(note_id_and_args: Option<Vec<NoteIdAndArgs>>) -> NoteIdAndArgsArray {
-        let note_id_and_args = note_id_and_args.unwrap_or_default();
-        NoteIdAndArgsArray(note_id_and_args)
-    }
-
-    pub fn push(&mut self, note_id_and_args: &NoteIdAndArgs) {
-        self.0.push(note_id_and_args.clone());
-    }
-}
-
 impl From<NoteIdAndArgsArray> for Vec<(NativeNoteId, Option<NativeNoteArgs>)> {
     fn from(note_id_and_args_array: NoteIdAndArgsArray) -> Self {
-        note_id_and_args_array.0.into_iter().map(Into::into).collect()
+        note_id_and_args_array.__inner.into_iter().map(Into::into).collect()
     }
 }
 
 impl From<&NoteIdAndArgsArray> for Vec<(NativeNoteId, Option<NativeNoteArgs>)> {
     fn from(note_id_and_args_array: &NoteIdAndArgsArray) -> Self {
-        note_id_and_args_array.0.iter().map(Into::into).collect()
+        note_id_and_args_array.__inner.iter().map(Into::into).collect()
     }
 }